+++ frontend/homepage/src/components/MiddleSections.tsx
@@ -1,0 +1,5 @@
+import middleHtml from "../content-middle.html?raw";
+
+export default function MiddleSections() {
+  return <div dangerouslySetInnerHTML={{ __html: middleHtml }} />;
+}